requirements.txt
Django
django-simple-menu
djangorestframework
